mcp_email_agent/requirements.txt
aiohttp==3.11.16
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
boto3==1.35.97
botocore==1.35.97
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jmespath==1.0.1
mcp==1.2.0
msal==1.28.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.45.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
